requirements.txt
flask
requests
generativeai
genai
python-dotenv
oauth2client